extra class for PersistenceView
